tests/wpt/tests/navigation-api/navigate-event/cross-window/submit-crossdocument-crossorigin.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<body>

<script>
async_test(t => {
  const url = new URL("resources/cross-origin-iframe-helper.html", location.href);
  url.hostname = get_host_info().REMOTE_HOST;
  const iframe = document.createElement("iframe")
  iframe.src = url;
  iframe.name = "windowname";
  document.body.append(iframe);

  const form = document.createElement("form");
  form.action = url;
  form.target = iframe.name;
  form.innerHTML = `<input type="hidden" name="postMessage-top-when-done" value="yes">`;
  document.body.append(form);

  window.onmessage = t.step_func_done(e => {
    // If we hit onnavigate in the target window, we'll get a different message, and fail.
    assert_equals(e.data, "DONE");
  });

  navigation.onnavigate = t.unreached_func("onnavigate must not fire in the source window");
  window.onload = t.step_func(() => form.submit());
}, "submitting a <form> element that navigates cross-document targeting a cross-origin window");
</script>
